Inital commit.
The `Dockerfile` for a base image form which all ADAPT services and
applications can inherit.  The intention is that it provide basic
services that will be available to everything that builds upon it.  For
example, in future this image might include a credentials service from
which any ADAPT application can safely retrieve secret data.

This image builds on the Alpine Linux image, which is *much* smaller
than the stock Ubuntu 16.04 LTS image (5MB vs 175MB).  This makes our
container own builds much lighter.  As it is based on Busybox, Alpine's
utilities are not as fully featured as Ubuntu's, but this extra
functionality will rarely, if ever, be used when running applications.
Alpine's philosophy aligns nicely with Docker's recommendation of
treating containers as app bundles, not servers.

An obvious benefit of Ubuntu is that it is a consistent and stable
platform that has become a de facto standard.  However, Hacker News
reports that Docker themselves have are standardising on Alpine Linux,
which suggests that the platform will be viable in the long term.